+++ src/S_SMP_ADSR.cpp
@@ -1,4 +1,4 @@
-// A Sony SPC700 chip (from Nintendo SNES) emulator module.
+// An envelope generator module based on the S-SMP chip from Nintendo SNES.
 // Copyright 2020 Christian Kauten
 //
 // Author: Christian Kauten (kautenja@auburn.edu)
@@ -24,7 +24,7 @@
 // MARK: Module
 // ---------------------------------------------------------------------------
 
-/// A Sony S-DSP chip (from Nintendo SNES) emulator module.
+/// An envelope generator module based on the S-SMP chip from Nintendo SNES.
 struct ChipS_SMP_ADSR : Module {
     /// the number of processing lanes on the module
     static constexpr unsigned LANES = 2;
@@ -224,13 +224,13 @@ struct ChipS_SMP_ADSRWidget : ModuleWidget {
     ///
     explicit ChipS_SMP_ADSRWidget(ChipS_SMP_ADSR *module) {
         setModule(module);
-        static constexpr auto panel = "res/S-SMP-ADSR.svg";
+        static constexpr auto panel = "res/S-SMP-ADSR-Light.svg";
         setPanel(APP->window->loadSvg(asset::plugin(plugin_instance, panel)));
         // panel screws
-        addChild(createWidget<ScrewBlack>(Vec(RACK_GRID_WIDTH, 0)));
-        addChild(createWidget<ScrewBlack>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, 0)));
-        addChild(createWidget<ScrewBlack>(Vec(RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
-        addChild(createWidget<ScrewBlack>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
+        addChild(createWidget<ScrewSilver>(Vec(RACK_GRID_WIDTH, 0)));
+        addChild(createWidget<ScrewSilver>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, 0)));
+        addChild(createWidget<ScrewSilver>(Vec(RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
+        addChild(createWidget<ScrewSilver>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
         for (unsigned i = 0; i < ChipS_SMP_ADSR::LANES; i++) {
             // Gate, Retrig, Output
             addInput(createInput<PJ301MPort>(Vec(20, 45 + 169 * i), module, ChipS_SMP_ADSR::INPUT_GATE + i));