+++ src/plugin.cpp
@@ -36,8 +36,10 @@ void init(Plugin* instance) {
     instance->addModel(modelChipPOKEY);
     instance->addModel(modelChipS_SMP);
     instance->addModel(modelChipS_SMP_ADSR);
+    instance->addModel(modelChipS_SMP_BRR);
     instance->addModel(modelChipS_SMP_Echo);
     instance->addModel(modelChipS_SMP_Gauss);
+    instance->addModel(modelChipS_SMP_Blank);
     instance->addModel(modelChipSCC);
     instance->addModel(modelChipSN76489);
     instance->addModel(modelChipTurboGrafx16);