+++ CHANGELOG.md
@@ -170,3 +170,7 @@
 -   new module: S-SMP(ADSR)
     -   Envelope generator from the SNES
     -   Attack, Decay, Sustain, and Release stages
+-   new module: S-SMP(BRR)
+    -   Bit-Rate Reduction (BRR) based sampler/sample player
+-   new module: S-SMP(Blank)
+    -   a blank panel with an illustration of the S-SMP audio processing chip